Extract Pending Withdrawals to separate component
